[CP-beta]Fix `getOffsetForCaret` crash (#146813)
This pull request is created by [automatic cherry pick workflow](https://github.com/flutter/flutter/wiki/Flutter-Cherrypick-Process#automatically-creates-a-cherry-pick-request)
Please fill in the form below, and a flutter domain expert will evaluate this cherry pick request.

What is the link to the issue this cherry-pick is addressing?

b/333560406. Couldn't find a github issue for it

Explain this cherry pick in one line that is accessible to most Flutter developers. See [best practices](https://github.com/flutter/flutter/wiki/Hotfix-Documentation-Best-Practices) for examples

Fixes a crash that may happen when the user enters certain characters in a text field.

What is the impact (ex. visual jank on Samsung phones, app crash, cannot ship an iOS app)? Does it impact development (ex. flutter doctor crashes when Android Studio is installed), or the shipping production app (the app crashes on launch)

Production apps crash when the user enters certain characters in certain conditions.

b/333560406 has some related stats.

Is there a workaround for this issue?

No

What is the risk level of this cherry-pick?

Are you confident that your fix is well-tested by automated tests?

What are the steps to validate that this fix works?

Run the test in the PR. The issue didn't come with a real world repro.